requirements.txt
snowflake-snowpark-python[pandas]
snowflake-cli-labs
python-dotenv
pytest
jinja2
pyyaml
boto3